refactor(e2e): move common/kind to internal/kind
Verbatim move of the package and its tests, plus the two import paths.
No behaviour change.

internal/ keeps these helpers out of reach of the root module and pkg/sdk,
and gives go test an exact selector for them.

make test selected ./common/... , which no longer contains the package, so
the target now selects ./internal/... as well. Without that the ten kind
tests would stop running and the target would still exit 0.
